requirements.txt
alembic==1.7.6
asttokens==2.0.5
backcall==0.2.0
bcolz-zipline==1.2.4
black==22.1.0
Bottleneck==1.3.2
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.1
empyrical-reloaded==0.5.8
entrypoints==0.4
exchange-calendars==3.3
executing==0.8.2
fonttools==4.29.1
greenlet==1.1.2
h5py==3.6.0
idna==3.3
importlib-metadata==4.11.1
importlib-resources==5.4.0
intervaltree==3.1.0
ipykernel==6.9.1
ipython==8.0.1
iso3166==2.0.2
iso4217==1.8.20211001
jedi==0.18.1
joblib==1.1.0
jupyter-client==7.1.2
jupyter-core==4.9.2
kiwisolver==1.3.2
korean-lunar-calendar==0.2.1
Logbook==1.5.3
lru-dict==1.1.7
lxml==4.8.0
Mako==1.1.6
MarkupSafe==2.1.0
matplotlib==3.5.1
matplotlib-inline==0.1.3
multipledispatch==0.6.0
multitasking==0.0.10
mypy-extensions==0.4.3
nest-asyncio==1.5.4
networkx==2.6.3
numexpr==2.8.1
numpy==1.22.2
packaging==21.3
pandas==1.4.1
pandas-datareader==0.10.0
parso==0.8.3
pathspec==0.9.0
patsy==0.5.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
platformdirs==2.5.1
prompt-toolkit==3.0.28
ptyprocess==0.7.0
pure-eval==0.2.2
pyfolio-reloaded==0.9.4
Pygments==2.11.2
pyluach==1.3.0
pyparsing==3.0.7
python-dateutil==2.8.2
python-interface==1.6.1
pytz==2021.3
pyzmq==22.3.0
requests==2.27.1
scikit-learn==1.0.2
scipy==1.8.0
seaborn==0.11.2
six==1.16.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.31
stack-data==0.2.0
statsmodels==0.13.2
TA-Lib==0.4.24
tables==3.7.0
threadpoolctl==3.1.0
tomli==2.0.1
toolz==0.11.2
tornado==6.1
trading-calendars==2.1.1
traitlets==5.1.1
typing-extensions==4.1.1
urllib3==1.26.8
wcwidth==0.2.5
yfinance==0.1.70
zipline-reloaded==2.2.0
zipp==3.7.0
